feat: Image annotator improvements (#1438)

+++ py/h2o_wave/ui.py
@@ -2362,10 +2362,10 @@ def image_annotator_tag(
 
 
 def image_annotator_rect(
-        x1: int,
-        y1: int,
-        x2: int,
-        y2: int,
+        x1: float,
+        y1: float,
+        x2: float,
+        y2: float,
 ) -> ImageAnnotatorShape:
     """Create a rectangular annotation shape.
